--- MusicApp/app/src/main/java/com/hoaiphong/composeui/ui/screen/myinfo/AvatarImage.kt
@@ -1,55 +1,0 @@
-package com.hoaiphong.composeui.ui.screen.myinfo
-
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import coil.Coil
-import com.hoaiphong.composeui.R
-
-
-@Composable
-fun AvatarImage(avatarPath: String?, modifier: Modifier = Modifier) {
-    Box(
-        modifier = modifier
-            .fillMaxSize()
-            .clip(CircleShape)
-            .border(2.dp, MaterialTheme.colorScheme.primary, CircleShape)
-    ) {
-        val avatarBitmap = remember(avatarPath) {
-            avatarPath?.let {
-                BitmapFactory.decodeFile(it)?.let { bmp ->
-                    Bitmap.createScaledBitmap(bmp, 300, 300, true)
-                }
-            }
-        }
-
-        if (avatarBitmap != null) {
-            Image(
-                bitmap = avatarBitmap.asImageBitmap(),
-                contentDescription = "User Avatar",
-                modifier = Modifier.fillMaxSize(),
-                contentScale = ContentScale.Crop
-            )
-        } else {
-            Image(
-                painter = painterResource(id = R.drawable.avata),
-                contentDescription = "Default Avatar",
-                modifier = Modifier.fillMaxSize(),
-                contentScale = ContentScale.Crop
-            )
-        }
-    }
-}
