--- MusicApp/app/src/main/java/com/hoaiphong/composeui/ui/screen/myinfo/rememberImagePicker.kt
@@ -1,19 +1,0 @@
-package com.hoaiphong.composeui.ui.screen.myinfo
-
-import android.content.Context
-import android.net.Uri
-import androidx.activity.compose.ManagedActivityResultLauncher
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.runtime.Composable
-
-@Composable
-fun rememberImagePicker(
-    context: Context,
-    onImagePicked: (Uri?) -> Unit
-): ManagedActivityResultLauncher<String, Uri?> {
-    return rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.GetContent(),
-        onResult = onImagePicked
-    )
-}
